// frontend/src/app/interfaces/ToDo.ts
export interface ToDoInterface {
  id: string;
  userId: string;
  title: string;
  description: string;
  checkBoxList: any[];
  startTime: string;
  endTime: string;
  estimatedTime: number;
  progress: number;
  priority: number;
  desireToDo: number;
  avodiance: number;
  f: number;
  g: number;
  totalPoint: number;
  positionX: number;
  positionY: number;
  width: number;
  height: number;
  area: number;
  color: string;
}